Fix the threaded-sweep data race, GPU/CPU precision drift, test gates, and CLI grammar
cpu_lib: threads no longer read a neighbor slab's boundary slice while
another thread writes it. Each thread skips the first slice of its
slab. The skipped slices are swept serially after the join. The result
is deterministic for a fixed thread count.

gpu_lib: fmin3 and fmax3 now take and return double. The float versions
truncated the double grid coordinates before the ceil and floor that
select the intersection band. That truncation could flip inside and
outside signs and defeated the --fmad=false CPU parity. Verified on an
RTX 4090: zero sign mismatches.

tests: the comparison gates can now fail. Zero sign mismatches are
required. Near-band differences are capped at dx/2. The far-field cap
remains because the CPU sweep and the GPU Jacobi iteration legitimately
diverge far from the surface. test_correctness reuses the shared
compare_sdf_grids instead of a private copy. The CLI tests are updated
from the old grammar they had drifted against. The drift surfaced once
the tests ran the freshly built binary instead of a stale one. The
tests/resources gitignore exception is narrowed so test outputs stay
ignored.

app: remove the magic threshold that treated a positional value below
20 as padding. The argument count now disambiguates. Honor the
advertised fourth positional padding value. Document -t, -p, and the
hardware auto-detection in the help text. Rename mode_precise to
use_grid_dimensions. Use snprintf and guard a division by zero.

Fix all 16 /W4 findings: unused parameters, size_t narrowing, and a
tolower cast. The build is warning-clean on MSVC 19.51 with /W4 and
/permissive-.

tools submodule: detect Visual Studio 2026 and BuildTools editions, and
update the submodule README.

Verified: 15 of 15 ctest suites pass on VS 2026 with CUDA 13.3 (GPU)
and CPU-only.

+++ gpu_lib/makelevelset3_gpu.cu
@@ -78,25 +78,32 @@ __device__ inline DistTriPair unpack_dist_tri(unsigned long long packed) {
 }
 
 /**
- * @brief Compute minimum of three floats
+ * @brief Compute minimum of three values
+ *
+ * Double precision on purpose: the callers pass double grid coordinates,
+ * and the CPU implementation performs this computation in double. A float
+ * version would truncate before the ceil/floor that select the
+ * intersection band and could flip inside/outside signs on
+ * large-coordinate meshes.
+ *
  * @param a First value
  * @param b Second value
  * @param c Third value
  * @return Minimum of a, b, c
  */
-__device__ inline float fmin3(float a, float b, float c) {
-    return fminf(fminf(a, b), c);
+__device__ inline double fmin3(double a, double b, double c) {
+    return fmin(fmin(a, b), c);
 }
 
 /**
- * @brief Compute maximum of three floats
+ * @brief Compute maximum of three values (double precision, see fmin3)
  * @param a First value
  * @param b Second value
  * @param c Third value
  * @return Maximum of a, b, c
  */
-__device__ inline float fmax3(float a, float b, float c) {
-    return fmaxf(fmaxf(a, b), c);
+__device__ inline double fmax3(double a, double b, double c) {
+    return fmax(fmax(a, b), c);
 }
 
 /**
